refactor: 상대경로 -> 절대경로 수정
