+++ src/app/(pages)/(Home)/explore/page.tsx
@@ -1,4 +1,4 @@
-import ExplorePage from "@/app/_components/pages/ExplorePage";
+import ExplorePage from "@components/pages/ExplorePage";
 import React from "react";
 
 export default function page() {
