+++ src/app/(pages)/(Home)/layout.tsx
@@ -1,6 +1,6 @@
 "use client";
-import Gnb from "@/app/_components/organism/Gnb/Gnb";
-import { HomeLayoutStyle } from "@/app/_styles/common.style";
+import Gnb from "@components/organism/Gnb/Gnb";
+import { HomeLayoutStyle } from "@/_styles/common.style";
 
 export default function HomeLayout({
   children,
