+++ src/app/(pages)/(Home)/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import HomePage from "@/app/_components/pages/HomePage";
+import HomePage from "@components/pages/HomePage";
 
 export default function Home() {
   return <HomePage />;
