+++ src/app/(pages)/login/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import LoginPage from "@/app/_components/pages/LoginPage";
+import LoginPage from "@components/pages/LoginPage";
 import React from "react";
 
 export default function page() {
