+++ src/app/_components/molecules/ModalContent/editModal/EditImageModal.tsx
@@ -1,12 +1,12 @@
-import { ModalStyle } from "@/app/_components/atoms/modal/Modal.style";
+import { ModalStyle } from "@components/atoms/modal/Modal.style";
 import React from "react";
 import { Content } from "./EditImageModal.style";
-import BaseButton from "@/app/_components/atoms/button/BaseButton";
+import BaseButton from "@components/atoms/button/BaseButton";
 import { faArrowLeft } from "@fortawesome/free-solid-svg-icons";
-import useModalStore from "@/app/_stores/client/modalStore";
+import useModalStore from "@/_stores/client/modalStore";
 import Image from "next/image";
-import IconButton from "@/app/_components/atoms/button/IconButton";
-import { MODAL_NAME } from "@/app/_constant/modal";
+import IconButton from "@components/atoms/button/IconButton";
+import { MODAL_NAME } from "@/_constant/modal";
 
 export default function EditImageModal() {
   const { selectedImage, setModal } = useModalStore();
