+++ src/app/_components/molecules/photoSet/PhotoSet.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Container } from "./PhotoSet.style";
-import PhotoPiece, { PhotoPieceProps } from "../../atoms/photoPiece/PhotoPiece";
+import PhotoPiece, { PhotoPieceProps } from "@components/atoms/photoPiece/PhotoPiece";
 
 export default function PhotoSet({ pieces }: { pieces: PhotoPieceProps[] }) {
   return (
