+++ src/app/_components/molecules/user/UswerSkeleton.style.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { UserStyle } from "./User.style";
-import { AvatarStyle } from "../../atoms/avatar/Avatar.style";
-import { SkeletonAni } from "@/app/_styles/common.style";
+import { AvatarStyle } from "@components/atoms/avatar/Avatar.style";
+import { SkeletonAni } from "@/_styles/common.style";
 
 export const Skeleton = {
   User: styled(UserStyle.User)``,
