+++ src/app/_components/organism/feedlist/FeedList.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Container } from "./FeedList.style";
-import Feed, { FeedProps } from "../../molecules/feed/Feed";
-import { useCustomQuery } from "@/app/_hooks/useFetch";
-import { QUERY_KEYS } from "@/app/_stores/server/queryKeys";
-import { FeedService } from "@/app/_services/feed_service";
+import Feed, { FeedProps } from "@components/molecules/feed/Feed";
+import { useCustomQuery } from "@/_hooks/useFetch";
+import { QUERY_KEYS } from "@/_stores/server/queryKeys";
+import { FeedService } from "@/_services/feed_service";
 import FeedSkeleton from "./FeedSkeleton";
 
 export default function FeedList() {
