+++ src/app/_components/organism/form/SignupForm.tsx
@@ -1,15 +1,15 @@
 import React, { ChangeEvent, useEffect, useState } from "react";
-import LabeledInput from "../../molecules/Input/LabeledInput";
-import { INPUT_TEXT } from "@/app/_constant/input";
-import BaseButton from "../../atoms/button/BaseButton";
+import LabeledInput from "@components/molecules/Input/LabeledInput";
+import { INPUT_TEXT } from "@/_constant/input";
+import BaseButton from "@components/atoms/button/BaseButton";
 import { Notice } from "./SignupForm.style";
-import { AuthService } from "@/app/_services/auth_service";
+import { AuthService } from "@/_services/auth_service";
 import { setCookie } from "cookies-next";
 import { useRouter } from "next/navigation";
-import useAuthStore from "@/app/_stores/client/authStore";
-import { authErrorHandler } from "@/app/_utils/authErrorHandler";
-import { Form } from "@/app/_styles/common.style";
-import { useCustomMutation } from "@/app/_hooks/useFetch";
+import useAuthStore from "@/_stores/client/authStore";
+import { authErrorHandler } from "@/_utils/authErrorHandler";
+import { Form } from "@/_styles/common.style";
+import { useCustomMutation } from "@/_hooks/useFetch";
 import { User } from "firebase/auth";
 const { EMAIL, NAME, USERNAME, PASSWORD } = INPUT_TEXT;
