+++ src/app/_components/organism/storyBar/StoryBar.style.tsx
@@ -1,4 +1,4 @@
-import { SkeletonAni } from "@/app/_styles/common.style";
+import { SkeletonAni } from "@/_styles/common.style";
 import styled from "styled-components";
 
 export const StoryStyle = {
