+++ src/app/_components/organism/storyBar/StoryBar.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { StoryStyle } from "./StoryBar.style";
-import User, { UserPorps } from "../../molecules/user/User";
-import { useCustomQuery } from "@/app/_hooks/useFetch";
-import { QUERY_KEYS } from "@/app/_stores/server/queryKeys";
-import { StoryService } from "@/app/_services/story_service";
+import User, { UserPorps } from "@components/molecules/user/User";
+import { useCustomQuery } from "@/_hooks/useFetch";
+import { QUERY_KEYS } from "@/_stores/server/queryKeys";
+import { StoryService } from "@/_services/story_service";
 import StorySkeleton from "./StorySkeleton";
 
 export default function StoryBar() {
