+++ src/app/_components/organism/storyBar/StorySkeleton.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Skeleton } from "./StorySkeleton.style";
-import UserSkeleton from "../../molecules/user/UserSkeleton";
+import UserSkeleton from "@components/molecules/user/UserSkeleton";
 
 export default function StorySkeleton() {
   const skeletonUsers = Array.from({ length: 8 }).map((_, idx) => <UserSkeleton key={idx} />);
