+++ src/app/_components/pages/ExplorePage.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import ExploreTemplate from "../template/explore/ExploreTemplate";
+import ExploreTemplate from "@components/template/explore/ExploreTemplate";
 
 export default function ExplorePage() {
   return <ExploreTemplate />;
