+++ src/app/_components/pages/HomePage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import HomeTemplate from "../template/home/HomeTemplate";
+import HomeTemplate from "@components/template/home/HomeTemplate";
 
 export default function HomePage() {
   return <HomeTemplate />;
