+++ src/app/_components/pages/SignupPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SignupTemplate from "../template/signup/SignupTemplate";
+import SignupTemplate from "@components/template/signup/SignupTemplate";
 
 export default function SignupPage() {
   return <SignupTemplate />;
