+++ src/app/_components/template/explore/ExploreTemplate.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { HomeLayoutStyle } from "@/app/_styles/common.style";
+import { HomeLayoutStyle } from "@/_styles/common.style";
 import { Template } from "./Template.style";
-import PhotoList from "../../organism/photolist/PhotoList";
+import PhotoList from "@components/organism/photolist/PhotoList";
 
 export default function ExploreTemplate() {
   return (
