+++ src/app/_components/template/explore/Template.style.tsx
@@ -1,4 +1,4 @@
-import { HomeLayoutStyle } from "@/app/_styles/common.style";
+import { HomeLayoutStyle } from "@/_styles/common.style";
 import styled from "styled-components";
 
 export const Template = {
