+++ src/app/_components/template/signup/SignupTemplate.tsx
@@ -1,6 +1,6 @@
-import { PageWrapper } from "@/app/_styles/common.style";
+import { PageWrapper } from "@/_styles/common.style";
 import React from "react";
-import SignupCard from "../../organism/group/SingupCard";
+import SignupCard from "@components/organism/group/SingupCard";
 
 export default function SignupTemplate() {
   return (
