+++ src/app/_dummyData/feedDummy.ts
@@ -1,6 +1,6 @@
 import { faker } from "@faker-js/faker";
 import { v4 as uuidv4 } from "uuid";
-import { getRandomBoolean, getRandomNumber } from "../_utils/utils";
+import { getRandomBoolean, getRandomNumber } from "@/_utils/utils";
 
 export function createFeeds(number: number) {
   const feedList = [];
